Fixed namespaces after Infrastructure move for Authentication https://github.com/achobanov/web-application/issues/3
